Remove size_t usage from memoc library benchmark

+++ benchmark/memoc/allocators.cpp
@@ -1,6 +1,5 @@
 #include <benchmark/benchmark.h>
 
-#include <cstddef>
 #include <cstdint>
 #include <vector>
 #include <memory>
@@ -8,11 +7,11 @@
 #include <memoc/allocators.h>
 
 struct Test_data {
-    std::vector<std::size_t> allocation_sizes{};
-    std::vector<std::size_t> choosen_size_indices{};
+    std::vector<std::int64_t> allocation_sizes{};
+    std::vector<std::int64_t> choosen_size_indices{};
 };
 
-template <std::size_t Min_allocation_size, std::size_t Max_allocation_size, std::size_t Number_of_allocations>
+template <std::int64_t Min_allocation_size, std::int64_t Max_allocation_size, std::int64_t Number_of_allocations>
 Test_data test_data()
 {
     static_assert(Min_allocation_size % 2 == 0);
@@ -21,11 +20,11 @@ Test_data test_data()
 
     Test_data td;
 
-    for (std::size_t i = Min_allocation_size; i <= Max_allocation_size; i *= 2) {
+    for (std::int64_t i = Min_allocation_size; i <= Max_allocation_size; i *= 2) {
         td.allocation_sizes.push_back(i);
     }
 
-    for (std::size_t i = 0; i < Number_of_allocations; ++i) {
+    for (std::int64_t i = 0; i < Number_of_allocations; ++i) {
         td.choosen_size_indices.push_back(i % td.allocation_sizes.size());
     }
 
@@ -38,7 +37,7 @@ static void BM_default_allocator(benchmark::State& state)
     auto td = test_data<16, 64, 64>();
 
     for (auto _ : state) {
-        for (std::size_t i : td.choosen_size_indices) {
+        for (std::int64_t i : td.choosen_size_indices) {
             auto p = alloc.allocate(td.allocation_sizes[i]);
             alloc.deallocate(p, td.allocation_sizes[i]);
         }
@@ -48,7 +47,7 @@ BENCHMARK(BM_default_allocator);
 
 template <class Allocator>
 void perform_allocations(Allocator* alloc, const Test_data& td) {
-    for (std::size_t i : td.choosen_size_indices) {
+    for (std::int64_t i : td.choosen_size_indices) {
         auto b = alloc->allocate(td.allocation_sizes[i]);
         alloc->deallocate(&b);
     }
@@ -128,10 +127,10 @@ static void BM_hybrid_allocator(benchmark::State& state)
 }
 BENCHMARK(BM_hybrid_allocator);
 
-template <class Allocator, typename T, std::size_t Number_of_allocations>
+template <class Allocator, typename T, std::int64_t Number_of_allocations>
 void perform_vector_allocations() {
     std::vector<T, Allocator> v{};
-    for (std::size_t i = 0; i < Number_of_allocations; ++i) {
+    for (std::int64_t i = 0; i < Number_of_allocations; ++i) {
         v.push_back(static_cast<T>(i));
     }
 }
